Move icons into components directory, away from general assets
